Added calls delegate
